forget float tensor instruction

+++ testing/test_model.py
@@ -5,6 +5,6 @@ import torch
 # testing countception model
 def test_countception():
     a = [[[[0 for i in range(10)] for i in range(10)]]]
-    inp = torch.tensor(a)
+    inp = torch.tensor(a).float()
     model = imed_model.Countception(in_channel=1, out_channel=1)
     model(inp)
